requirements.txt
# Required packages for LinkedIn Voice Assistant

# OpenAGI - AI automation
oagi

# Web framework
Flask>=3.0.0
flask-cors>=4.0.0

# Speech-to-Text
elevenlabs>=1.0.0

# Voice recognition and text-to-speech (for CLI version)
SpeechRecognition>=3.10.0
pyttsx3>=2.90
pyaudio>=0.2.13

# Google Sheets integration (optional)
gspread>=5.12.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
